Cleanup of bootstrap/loadout scripts
This changeset fixes a number of minor issues with my setup script: the
defaultsdiff script (which I am trying to use to track changes made from
stock macOS defaults) had some bashisms in it which caused problems when
run as a shell function; there were a few typos and logging messages
that could be removed; Slack and sphinx were missing from the initial
loadout.

It's worth noting that `mas` will not download from the Mac App Store
for an initial download (if the Apple ID on the computer has not
downloaded it before); there's probably something that can be done with
the "purchase" subcommand, since all Mac App Store applications in
current use are free. This isn't an issue for my personal account but
might cause confusion in the future.